[refactor] - components abstraction
