+++ src/components/BillCard/BillCard.jsx
@@ -1,7 +1,10 @@
-import { useAuth } from "../../../context/auth-context";
+import { useAuth } from "../../context/auth-context";
+import { useNavigate } from "react-router-dom";
 
 const BillCard = () => {
   const { userState } = useAuth();
+
+  const navigate = useNavigate();
 
   const initialBill = {
     totalProductPrice: 0,
@@ -52,7 +55,10 @@ const BillCard = () => {
           on this order
         </p>
       </div>
-      <button className="btn btn-primary btn-lg">
+      <button
+        className="btn btn-accent-light btn-lg"
+        onClick={() => navigate("/checkout")}
+      >
         <i className="fas fa-shopping-bag" />
         Place order
       </button>
