+++ src/pages/Cart/cart.js
@@ -1,5 +1,5 @@
-import { CartProductCards } from "./cart-components/cart-product-card";
-import { BillCard } from "./cart-components/bill-card";
+import { CartProductCards } from "../../components/CartProductCard/CartProductCard";
+import { BillCard } from "../../components/BillCard/BillCard";
 import { useAuth } from "../../context/auth-context";
 import { EmptyBucket } from "../../components/EmptyBucket/EmptyBucket";
